refactor: move brc workflow services into sites/brc-analytics (#1637)
