+++ sites/brc-analytics/views/OrganismView/hooks/UseShowPangenome/hook.ts
@@ -1,5 +1,5 @@
-import { getPangenome } from "@/services/workflows/brc/entities";
 import type { Pangenome } from "@brc/apis/pangenome";
+import { getPangenome } from "@brc/services/workflows/entities";
 import { useFeatureFlag } from "@databiosphere/findable-ui/lib/hooks/useFeatureFlag/useFeatureFlag";
 
 /**
